Make Chrome window fill Chrome session on start
Previously Chrome would start at the size it last had at if it
had been opened previously, and it was not possible to resize the window
as the Chrome session type does not have a window manager.

This is particularly needed now that Clusterware users can specify a
session geometry on the command line, but could still have caused issues
previously if Chrome was opened and resized within a session of another
type and then a Chrome session was started.